Db

Description

This module handles requests; fetch or execution of queries; to the database. Such as inserting new tweets from Twitter into the local database. All statements are prepared to prevent SQL injection.

Documentation

data DbConfig

DbConfig contains info needed to connect to MySQL such as user, password, etc.

Constructors

DbConfig

Fields

dbName :: String

dbUser :: String

dbPassword :: String

newConn :: DbConfig -> IO Connection

Creates a Databse connection through the configuration file applicaiton.conf

fetchT :: (QueryResults r, QueryParams q) => Pool Connection -> q -> Query -> IO [r]

Fetch command which accepts arguments

fetchSimpleT :: QueryResults r => Pool Connection -> Query -> IO [r]

Simple fetch without prepared arguments

execSqlT :: QueryParams q => Pool Connection -> q -> Query -> IO Int64

Execute with arguments to update database

findUserByLogin :: Pool Connection -> String -> IO (Maybe String)

Returns the encrypted password for a user from the DB, if user is found

listTweets :: Pool Connection -> Text -> IO (Maybe [Tweet])

Lists all tweets matching a certain screen_name, based on SQL LIKE; %% for all tweets

findTweet :: Pool Connection -> Text -> IO (Maybe Tweet)

List a single tweet based on user ID

insertTweets :: Pool Connection -> Maybe [Tweet] -> ActionT Text IO ()

Insert list of tweets into database based on returned tweets from Querying the REST API of twitter

insertTweet :: Pool Connection -> Tweet -> ActionT Text IO ()

Insert a single tweet into the database

deleteTweet :: Pool Connection -> Text -> IO (Maybe Integer)

Delete a tweet in the database based on id

listUsers :: Pool Connection -> Text -> IO (Maybe [User])

Lists all users matching a certain screen_name, based on SQL LIKE; %% for all tweets

insertUsers :: Pool Connection -> Maybe [User] -> ActionT Text IO ()

Insert list of users into database based on returned users from Querying the REST API of twitter

insertUser :: Pool Connection -> User -> ActionT Text IO ()

Insert a single user into the database

updateUser

Arguments

:: Pool Connection

-> Maybe User
Was the JSON body in correct form?

-> ActionT Text IO ()

Update a user based on unique screen_name