Views

Description

This module handles the rendering of haskell type data and encodes it into JSON format for display to the end user either as response to a cURL request or browser request.

Documentation

listedTweets :: Maybe [Tweet] -> ActionM ()

Takes a list of tweets and renders it to json for display to the request holder

viewTweet :: Maybe Tweet -> ActionM ()

Takes a Maybe Tweet. And renders it to JSON

deletedTweet :: Maybe Integer -> ActionM ()

Displays message for deleting a tweet.

listedUsers :: Maybe [User] -> ActionM ()

Takes a Maybe list of users and renders it to json for display to the request holder

updatedUser :: Maybe User -> ActionM ()

Takes a Maybe user and renders it to json for display to the request holder. Error thrown if PUT request was erroneous

viewUser :: Maybe User -> ActionM ()

Takes a Maybe user and does same as above however the error here is thrown if user not found